requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
nltk==3.9.1
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
schedule==1.2.2
textblob==0.19.0
tqdm==4.67.1
urllib3==2.4.0
Werkzeug==3.1.3
